refac(91026): Extrai task gerar_previa_demonstrativo_financeiro_async
