+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -2,3 +2,5 @@ from .tasks_legado import *
 
 from .concluir_prestacao_de_contas import concluir_prestacao_de_contas_async
 from .processa_carga import processa_carga_async
+from .gerar_previa_demonstrativo_financeiro import gerar_previa_demonstrativo_financeiro_async
+
